requirements.txt
django==1.5.1
djangorestframework==2.2.6
